pyproject.toml
[project]
name = "whisp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "espeakng-loader>=0.2.4",
    "gguf>=0.18.0",
    "librosa>=0.11.0",
    "onnx>=1.20.0",
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.6",
    "phonemizer-fork>=3.3.2",
    "safetensors>=0.5.3",
    "snac>=1.2.1",
    "tensorboard>=2.20.0",
    "torch>=2.11.0",
    "tqdm>=4.67.3",
    "transformers>=5.7.0",
]
